[#258] delete the engine's timeoutChecker scheduler and merge it to sessionManager's scheduler fix #258 Squashed commit of the following:
commit 5b0a4d1f77b32cef00f55b80b8562752c1f462d7
Author: zen <xinjingziranchan@gmail.com>
Date:   Wed Dec 16 10:15:16 2020 +0800

    fix closing session log

commit c0fe24d06638f9be18d91a4ad05943a023967c62
Author: zen <xinjingziranchan@gmail.com>
Date:   Sat Dec 12 22:08:25 2020 +0800

    move this check task to SparkSQLSessionManager.

commit 1968913702872f0ba433679596596f114f2e16d1
Author: zen <xinjingziranchan@gmail.com>
Date:   Thu Dec 10 15:55:32 2020 +0800

    add log when close session

commit 26a944787c63666813a0d64e67a36b671640d493
Author: zen <xinjingziranchan@gmail.com>
Date:   Thu Dec 10 11:32:45 2020 +0800

    delete the engine's timeoutChecker scheduler and merge it to  sessionManager's scheduler.
